fix reflection mirror problems
before, i was using the mirror from
scala.reflection.runtime.currentMirror, which would give me the mirror
that was used to load longevity/emblem libraries. this has been shown to
be an error, we should be using the users mirror. so ive changed things
around to use the mirror that comes with the TypeTag (that is wrapped in
a TypeKey) provided by the user code.

additionally, from time to time we construct TypeTags (TypeKeys) from
the user's types, combining them with other types such as `Prop` or
`KeyVal`. the TypeTags we construct this way need to modified to use the
mirror of the user's types. See the pscan/kscan/iscan methods in PType
for an example of this.