specify and clarify corner case with @persistent macro
for @persistent, @polyPersistent, and @derivedPersistent, the old behavior was to just throw out the
first type that and existing companion object extended. if the user does not explicitly extend a
class, this first type will be scala.AnyRef. if they do explicitly extend a class, this first type
will be the class they extend. (note we are talking about classes here and not traits. while the
companion object can inherit from many traits, it can only inherit from a single class.)

this is strange behavior, and also inconsistent with the content of the scaladocs. behavior is
updated as follows:

- if existing companion object does not explicitly extend any class, or explicitly extends
  scala.AnyRef, the augmented companion object will extend the appropriate PType
- if existing companion object extends the appropriate PType, the augmented companion object will
  extends the same things the existing companion object extends
- if existing companion object extends any other class, then compiler error

the scaladocs have been updated to reflect actual behavior. tests cases also added where
pragmatic. (its pretty hard to write a test for the compiler error case.)