replace IO with Future in migrations
sucks but, silly me i expected IO was going to have some option to run tasks in parallel. at this
point i should really turn to Monix of FS2. but this would mean learning a new library and adding a
new longevity effect class, and right now, i just want to turn this thing in, so i can move on to
the next thing in my life. so, lame me, going with scala futures. hey we choose our battles. (still
using cats and iteratee-io)